+++ Tests/Integration/Microsoft365DSC/M365DSCIntegration.EXO.Create.Tests.ps1
@@ -885,6 +885,21 @@
                     TenantId              = $TenantId
                     CertificateThumbprint = $CertificateThumbprint
                 }
+                EXOSweepRule 'MyRule'
+                {
+                    ApplicationId         = $ApplicationId;
+                    CertificateThumbprint = $CertificateThumbprint;
+                    DestinationFolder     = "Test2:\Deleted Items";
+                    Enabled               = $True;
+                    Ensure                = "Present";
+                    KeepLatest            = 11;
+                    Mailbox               = "Test2";
+                    Name                  = "From Michelle";
+                    Provider              = "Exchange16";
+                    SenderName            = "michelle@fabrikam.com";
+                    SourceFolder          = "Test2:\Inbox";
+                    TenantId              = $TenantId;
+                }
                 EXOTransportRule 'ConfigureTransportRule'
                 {
                     Name                                          = "Ethical Wall - Sales and Executives Departments"