+++ Tests/Unit/Stubs/Microsoft365.psm1
@@ -1,4 +1,50 @@
 # region ExchangeOnlineManagement
+function Get-MailboxIRMAccess
+{
+    [CmdletBinding()]
+    param(
+        [Parameter()]
+        [System.Object]
+        $User,
+
+        [Parameter()]
+        [System.Object]
+        $Identity
+    )
+}
+
+function Set-MailboxIRMAccess
+{
+    [CmdletBinding()]
+    param(
+        [Parameter()]
+        [System.Object]
+        $User,
+
+        [Parameter()]
+        [System.Object]
+        $Identity,
+
+        [Parameter()]
+        [System.Object]
+        $AccessLevel
+    )
+}
+
+function Remove-MailboxIRMAccess
+{
+    [CmdletBinding()]
+    param(
+        [Parameter()]
+        [System.Object]
+        $User,
+
+        [Parameter()]
+        [System.Object]
+        $Identity
+    )
+}
+
 function Get-SweepRule
 {
     [CmdletBinding()]