+++ CHANGELOG.md
@@ -2,7 +2,11 @@
 
 # UNRELEASED
 
+* EXOOwaMailboxPolicy
+  * Add support for AccountTransferEnabled parameter
 * EXOSweepRule
+  * Initial Release.
+* FabricAdminTenantSettings
   * Initial Release.
 * IntuneDeviceControlPolicyWindows10
   * Initial Release
@@ -12,6 +16,8 @@
     FIXES [#5000](https://github.com/microsoft/Microsoft365DSC/issues/5000)
   * Fixes an issue where the exported nested CIM instances had too many line breaks.
   * Fixes an issue where Settings Catalog properties were not correctly handled.
+* DEPENDENCIES
+  * Updated MSCloudLoginAssistant to version 1.1.20.
 
 # 1.24.828.1
 
@@ -71,8 +77,6 @@
 * M365DSCUtil
   * Fix `Compare-PSCustomObjectArrays` by allowing empty arrays as input
     FIXES [#4952](https://github.com/microsoft/Microsoft365DSC/issues/4952)
-* O365OrgSettings
-  * FIXES [#4741](https://github.com/microsoft/Microsoft365DSC/issues/4741)
 * MISC
   * Improve module updates and PowerShell Core support across the DSC
     resources.