+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change log for Microsoft365DSC
 
-# UNRELEASED
+# 1.24.904.1
 
 * EXOOwaMailboxPolicy
   * Add support for AccountTransferEnabled parameter